Makefile and documentation clean up
Corrections to the README for environment setup.

Added information about environment setup so the responsibility
for handling the virtual environments is more on the user.

Reworked the Makefile to be simpler.

Corrected a stale command line reference.

Clean up for documentation related building.

Corrected .PHONY so that make now correctly picks up all cases.
Also cleaned up the build command line for docs and added cleandocs.

Signed-off-by: Francesco Di Natale <dinatale3@llnl.gov>