This commit captures the work done to rework parameters.
The commits here are the implementation of a change to rework
parameters to construct studies based on the parameters that
are used instead of all parameters. The previous logic led to
the repetition of certain steps needlessly which is costly for
simulations that have expensive preamble steps. This implementation
also prevents unused parameters from affecting the number of runs.

Additions of methods to the object API for finding parameters.

Reworks to name based on used parameters.

Addition to carry the union of parent and step params.

Addition of workspace logic to parameterization.

Addition of debug logging.

Signed-off-by: Francesco Di Natale <dinatale3@llnl.gov>