Implementation of an initial state reporting user interface.
These commits rework the necessary internals to better support a
more user friendly way to access a study DAG's status. The current
implementation constructs a csv in the study directory, to which
the user can query via 'maestro -s <study output path>'.

Docstring updates for StepRecord to include new interface.

Tweaks to the state tracking of StepRecords.

Addition of file locking and csv output for statuses

Tweak to have the conductor use the write_status method.

Addition of filelock to dependencies.

Addition of filelock to dev dependencies.

Addition of a missing comma.

Start of detection of status.

Addition of a CLI to report the status of a running study.

Addition of the tabulate requirement.

Removal of debug print statements and added docstrings/comments

Correction of the locking mechanism for status checks.

Tweak to strip newlines from headers.

Docstring corrections

Fixed a bug that would cause step states to not be set.

Signed-off-by: Francesco Di Natale <dinatale3@llnl.gov>