Implementation of a ScriptAdapter for the IBM LSF Scheduler.
Initial implementation of an LSF adapter.

Addition of LSF to the interface factory.

Correction to time format LSF correction.

Tweak to correct for casting

Further tweaks to the LSFScriptAdapter

Adjustments to the states the LSF adapter can return.

More tweaks to LSF states and status checking.

Update to the batch setting docstring

Tweak to make wallclock time entries two digits

Bugfix to the previous commit.

Signed-off-by: Francesco Di Natale <frank.dinatale1988@gmail.com>