Introduces the ability to throttle the number of inflight jobs.
Addition of a queue for inflight jobs and a submission limit.

Additions to check for submission throttles when executing ready steps.

Removal of name as an argument to _execute_record (record contains name already).

Addition of throttling to the CLI and Study staging.

Made clear in throttle help that 0 is a non-throttle value.

Missed formatter for throttle error.

Tweak to make the throttle conditional clearer.

Addition of submission parameter error checking in ExecutionGraph.

Correction of the ArgumentError to include greater than 0.

Signed-off-by: Francesco Di Natale <frank.dinatale1988@gmail.com>

Corrected submission attempt minimum to 1