Adds support for enhanced launcher token notations.
This PR continues to support the legacy notation of $(LAUNCHER)[n,p],
where n is the number of nodes a command requires and p is the number
of processors. This notation is expanded to loosen the requirement for
nodes (except in the legacy notation above) and changes the notation to
allow for more flexibility (nodes and processors can be notated out of
the above order). The new notations becomes $(LAUNCHER)[#n, #p] where
'#n' is not required ($(LAUNCHER)[#p] is also valid). See the study in
samples/documentation/launcher_tokens.yaml.

Initial logic and new regex.

Logic for parsing out launcher tokens.

Addition of a case for when launcher token appears on its own.

Update to the SLURM adapter to supposed unset node counts.

Added launcher token documentation study file.

Bug fixes for regex matching and processing.

Signed-off-by: Francesco Di Natale <dinatale3@llnl.gov>

Addition of error checking for allocations.