Exposed the restart limit settings to the CLI.
The restart limit controls the number of times max
a workflow step should restart.

Addition of a limit set to 0 as unlimited restarting.

Addition of restart limit to command line and CLI cleanup.

Added rlimit to the parameters passed to setup.

Addition of an rlimit check for zero or greater.

Signed-off-by: Francesco Di Natale <dinatale3@llnl.gov>