Addition of a feature that dumps generate scripts to /tmp
Correction of some docstring warnings.

Removal of script generation.

Addition of script generation to StepRecord.

Update to use the StepRecord script generation.

Refactoring to move launching to StepRecord.

Tweaks and bug fixes.

Tweak to make tmp_dir use a boolean.

Reworking of tmp_dir in ExecutionGraph.

Two versions of lulesh_sample1 for Unix and Mac.

Correction of flake8 warnings.

Addition of --usetmp to CLI.

Addition of use_tmp to Study setup.

Addition of logging and cleanup.

Addition of logging.

Addition of clean up after study completes.

Tweak to generate workspaces on the fly.

Signed-off-by: Francesco Di Natale <dinatale3@llnl.gov>