Test/interfaces/lsf (#215)
* Implementation of a ScriptAdapter for the IBM LSF Scheduler.

Initial implementation of an LSF adapter.

Addition of LSF to the interface factory.

Correction to time format LSF correction.

Tweak to correct for casting

Further tweaks to the LSFScriptAdapter

Adjustments to the states the LSF adapter can return.

More tweaks to LSF states and status checking.

Update to the batch setting docstring

Tweak to make wallclock time entries two digits

Bugfix to the previous commit.

Signed-off-by: Francesco Di Natale <frank.dinatale1988@gmail.com>

* Addition of GPU support.

* Addition of a cancel method.

* Addition of reservation submissions.

* Tweak to use the -nrs flag for jsrun.

* Changes to resource allocation parameters.

* Removal of some batch headers for LSF.

* Tweak to error code for NOJOBS status.

* Tweak to skip lines that are part of prev status line.

* Correction of --nrs

* Correction of task batch key to nodes.

* Tweaks to _substitute_parallel_command.

Now only pass in step.run by copy and append the popped keys as
step resources "snodes" and "sprocs".

* Correction to LSF adapter to correct node specification.

* Tweaks to checking status of LSF jobs.

A tweak to formatting of the output for bjobs. With the new format
we get termination reason, which allows us to check for a timed
out status.

* Correction to bjobs formatting and nojob check.

* Corrections to how nodes and procs are being passed.

* Correction of the bkill command with multiple job ids.

* Tweaks to check_jobs for LSF adapter.

* Implementation of a ScriptAdapter for the IBM LSF Scheduler.

Initial implementation of an LSF adapter.

Addition of LSF to the interface factory.

Correction to time format LSF correction.

Tweak to correct for casting

Further tweaks to the LSFScriptAdapter

Adjustments to the states the LSF adapter can return.

More tweaks to LSF states and status checking.

Update to the batch setting docstring

Tweak to make wallclock time entries two digits

Bugfix to the previous commit.

Signed-off-by: Francesco Di Natale <frank.dinatale1988@gmail.com>

* Addition of GPU support.

* Addition of a cancel method.

* Tweak to use the -nrs flag for jsrun.

* Changes to resource allocation parameters.

* Removal of some batch headers for LSF.

* Tweak to error code for NOJOBS status.

* Tweak to skip lines that are part of prev status line.

* Correction of task batch key to nodes.

* Tweaks to _substitute_parallel_command.

Now only pass in step.run by copy and append the popped keys as
step resources "snodes" and "sprocs".

* Correction to LSF adapter to correct node specification.

* Tweaks to checking status of LSF jobs.

A tweak to formatting of the output for bjobs. With the new format
we get termination reason, which allows us to check for a timed
out status.

* Correction to bjobs formatting and nojob check.

* Corrections to how nodes and procs are being passed.

* Correction of the bkill command with multiple job ids.

* Tweaks to check_jobs for LSF adapter.

* Addition of LSF to key for LSFScriptAdapter.

* Correction of lsf key.

* Addition of a debug statement to catch status command.

* Correction of bjobs command.

* Additions to status checks.

* Rearraging some debug logging.

* Testing to see if .split works.

* Further LSF tweaks.

* Revert back to split with strip.

* Removal of -q option due to excessive filtering.

* Correction of a missed merge

* Correction to use new Records structures.

* Style fix for line 207.

* Correction to SubmissionRecord creation.

* Decode output in check_status to enforce str type.

* Decode output in submit.

* Addition of retcode to a logger statement.