Add name and real_name StudyStep properties. (#314)
* Add name and real_name StudyStep properties.

There now needs to be an aliasing of nickname with a step name
when requested because adapters expect to use the name. StudyStep
objects now return their nickname if the nickname is set. Objects
such as the ExecutionGraph must now use the real_name for logistic
tracking, while adapters can continue to use name.

* Addition of extension property for Flux.

* Some typo and string style tweaks.