Limit to cryptography<3 for build (#353)
Short-term goal: The goal for now was to get TravisCI testing Python3.6 again. It was previously pulling source and trying to build cryptography at a more recent version from PyPi. This fix makes a wheel version available so that the recent change to include rust doesn't prevent testing.

Long-term: Eventually we will want to move to the most recent cryptography, but for now it breaks 3.6. The current specified schedule for 3.6 is that currently it is only receiving security updates until 12-2021 (when it will presumably be EOL). We will revisit this when we deprecate 3.6 support.