flux: make `force_broker` the default action
Problem: when creating maestro specs that need to be portable across
schedulers, it is assumed that a `${LAUNCHER}` call is needed in many
scripts, but when `force_broker` is `False`, this results in a doubling
up of the parallelism.  The call to the `submit` function within the
Flux adaptor launches multiple processes and then then `${LAUNCHER}`
call within the script gets expanded by the `parallelize` function to
also launch multiple processes.

Solution: make `force_broker` the default behavior to maximize
compatibilty with other schedulers.  If users want to avoid the extra
overhead of a nested Flux instance, they can always opt out of the
default behavior with `force_broker = False`, and then elide their use
of the `${LAUNCHER}` variable in the script.