Flux update (#359)
* flux0.26: add version checking during connection

Problem: there are several versions of the Flux adaptor and it is very
easy to use too new of an adaptor with too old of a Flux version (and
vice-versa)

Solution: check both the adaptor version and the Flux broker version
when making a connection via a new Flux handle and ensure the versions
match.  If the adaptor version is newer, log an error.  If the broker
version is newer, log a debug message letting the user know they might
benefit from choosing a newer adaptor.

* flux0.26: handle the case where the ngpus argument is a string

Problem: the `ngpus` argument to `submit` (as well as `parallelize`) can
sometimes be a string (e.g., when it isn't the default of 0). This
causes validation errors in the Flux jobspec class, which does runtime
type checking of the value

Solution: explicitly check if `ngpus` is a string, specifically a digit,
and if so, convert it to an `int` before passing to Flux

* flux0.26: fix number of slots for nested flux launches

Problem: when launching with multiple tasks in a nested Flux
instance (i.e., `force broker = True`), then the number of slots for the
nested instance should be set as the number of tasks not the number of
nodes

* flux: make `force_broker` the default action

Problem: when creating maestro specs that need to be portable across
schedulers, it is assumed that a `${LAUNCHER}` call is needed in many
scripts, but when `force_broker` is `False`, this results in a doubling
up of the parallelism.  The call to the `submit` function within the
Flux adaptor launches multiple processes and then then `${LAUNCHER}`
call within the script gets expanded by the `parallelize` function to
also launch multiple processes.

Solution: make `force_broker` the default behavior to maximize
compatibilty with other schedulers.  If users want to avoid the extra
overhead of a nested Flux instance, they can always opt out of the
default behavior with `force_broker = False`, and then elide their use
of the `${LAUNCHER}` variable in the script.

* Moved integer check to FluxScriptAdapter main class.

* Move the version check to FluxInterface base class.

* Update prior interfaces to use connect_to_flux method.

* Fix for an empty gpu string in a specification.

* Removal of gpu conversion in 0.26.0 Flux backend.

Co-authored-by: Francesco Di Natale <frank.dinatale1988@gmail.com>