Patch flux adapter to enable pickling to save study state and update it's resource/uri handling (#415)
Casts flux's JobID objects into string form to enable the dag to be pickled when conductor saves state. JobID objects are tied to flux's use of cffi and are not pickle-able themselves.

Base58 string form is used instead of the integer it's derived from so that logging is consistent with the view users have on the command line with flux reporting job id's in that format. This requires a separate re-instantiation of JobID objects upon waking to check the job status'.

Additional fixes:
* Improve the flexibility of resource specification in the flux adapter to align with the other adapters (i.e. not requiring nodes be specified).
* Change uri handling so env vars are not the only way to schedule jobs
* Updates to get working on flux only machines that require no bootstrapping or uri
  * Changes job submit call to be non-waitable by default on adapter 0.49 and newer as only owner of instance can do so (which is rarely the case on flux managed machines)
* Adds documentation and recipes for how to use Maestro with flux machines and flux instances running on non-flux scheduled machines
* Pins poetry version to 1.4.2 in CI to maintain support for python 3.7
* Fixes parsing of flux broker version to properly handle non-release versions of flux and logs that version into the flux batch script header comments and the maestro logs.